dada

+++ POS1/POS1/Main/Model/Inventory/InventoryModel.cs
@@ -10,9 +10,9 @@ namespace POS1.Main.Model.Inventory
     {
         public int ID { get; set; }
         public string NAME { get; set; }
-        public double? PRICE { get; set; }
+        public decimal? PRICE { get; set; }
         public int? QTY { get; set; }
-        public double? WHOLESALEPRICE { get; set; }
+        public decimal? WHOLESALEPRICE { get; set; }
 
         public List<InventoryModel> getAllInventory()
         {
